Bug #1423725 bsp: isp: Fix PC-lint warning
[root cause  ] format modify
[changes     ] isp_slice.h,isp_drv.c,isp_slice.c
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:9863
[test case]:capture preview
[reviewers   ] wei.ren
[change_type ] feature
[tag_product ] common

+++ core/isp2.6/inc/isp_slice.h
@@ -259,7 +259,7 @@ struct slice_ltm_map_info {
 	uint32_t tile_left_flag;
 	uint32_t tile_start_x;
 	uint32_t tile_start_y;
-	uint32_t mem_addr;
+	unsigned long mem_addr;
 };
 
 struct slice_noisefilter_info {
